modal_runner/pyproject.toml
[project]
name = "modal-runner"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "modal>=1.2.1",
    "pillow>=12.0.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "torchaudio>=2.9.0",
    "requests>=2.31.0", # For HTTPS proxy functionality
    "wandb", # For logging (optional)
    "gast>=0.5",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["inference_runner", "run_local", "inference"]
